Functional test for get_news_urls
This change introduces a logging feature to the webpage extractors,
which will only work in the case a top logger is created. This feature
is used in the test_extractor.py, which will create a log in tests/.

For now the log file will increment without limit. In future we should
erase the log file if any at the top of the test.